references/multi-agent-orchestration-main/apps/orchestrator_3_stream/backend/pyproject.toml
[project]
name = "orchestrator_3_stream_backend"
version = "1.0.0"
description = "Orchestrator 3 Stream - PostgreSQL-backed multi-agent orchestration"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "python-multipart",
    "asyncpg",
    "claude-agent-sdk",
    "rich",
    "pydantic",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = []
